refactor existing pulumi code
